Added initial version of Visual Basic 6 Proxy.
Added a method to copy between GPU textures.
Added a method to stop all sounds from a specific emitter.
Added Vector2 and Vector3 lerp functions.
Added rotation in the Sprite Renderer.
Reduced memory allocation for the Handle Allocator.
Fixed when creating a function that is only used as a framebuffer attachment.
Fixed deleting a pass object.